+++ source/displayfile.py
@@ -6,7 +6,8 @@ Módulo para o handler do display file.
 
 import gi
 
-from source.wireframe import Object
+from math import degrees
+from source.wireframe import Object, Window
 from source.file_system import FileSystem
 from source.transform import Vector
 
@@ -24,11 +25,13 @@ class DisplayFileHandler():
     objects: list[Object]
 
     # Atributos privados
+    _all_objects_normalized: bool
     _display_file_list: Gtk.ListStore
 
     def __init__(self, display_file_list: Gtk.ListStore, file_system: FileSystem) -> None:
 
         self.objects = []
+        self._all_objects_normalized = False
         self._display_file_list = display_file_list
 
     # Métodos
@@ -39,6 +42,7 @@ class DisplayFileHandler():
 
         self.objects.append(obj)
         self._display_file_list.append([obj.name, str(obj.position)])
+        self._all_objects_normalized = False
 
     # Por enquanto o id é o nome
     def remove_object(self, identification: str) -> None:
@@ -47,9 +51,9 @@ class DisplayFileHandler():
         '''
 
         for obj in self.objects:
-
             if obj.identification == identification:
                 self.objects.remove(obj)
+                self._all_objects_normalized = False
                 break
 
     def update_object_info(self, index: int) -> None:
@@ -67,14 +71,38 @@ class DisplayFileHandler():
         if len(self.objects) > 0:
             self.objects.pop()
             self._display_file_list.remove(self._display_file_list[-1].iter)
+            self._all_objects_normalized = False
 
     def save_world(self) -> None:
         '''
-        ...
+        Salva o mundo em um arquivo .obj.
         '''
 
-        # file_system.afafjka
+        raise NotImplementedError
 
-    def normalize_objects(self, window_center: Vector) -> None:
+    def request_normalization(self) -> None:
         '''
+        Define os objetos como não normalizados.
         '''
+
+        self._all_objects_normalized = False
+
+    def normalize_objects(self, window: Window) -> None:
+        '''
+        Normaliza todos os objetos.
+        '''
+
+        if not self._all_objects_normalized:
+            self._all_objects_normalized = True
+
+            window.normalized_translate(window.position * -1)
+            window.normalized_rotate(-window.rotation.z)
+
+            window_up = window.calculate_up_vector()
+            rotation = degrees(window_up * Vector(0.0, 1.0, 0.0))
+
+            if window_up.x > 0.0:
+                rotation = 360 - rotation
+
+            for obj in self.objects:
+                obj.normalize(window.position, rotation)