+++ source/editor.py
@@ -221,6 +221,37 @@ class EditorHandler():
                 self.update_spin_buttons()
                 self._temp_coords.clear()
 
+    def handle_key_press(self, key: str) -> None:
+        '''
+        Processa um evento de pressionamento de tecla.
+        '''
+
+        match key:
+            case 'q':
+                self._main_window.viewport_handler.rotate_window(-10)
+            case 'e':
+                self._main_window.viewport_handler.rotate_window(10)
+            case 'w':
+                self._main_window.viewport_handler.move_window(Vector(0.0, 10.0, 0.0))
+            case 'a':
+                self._main_window.viewport_handler.move_window(Vector(-10.0, 0.0, 0.0))
+            case 's':
+                self._main_window.viewport_handler.move_window(Vector(0.0, -10.0, 0.0))
+            case 'd':
+                self._main_window.viewport_handler.move_window(Vector(10.0, 0.0, 0.0))
+            case 'r':
+                self._main_window.viewport_handler.reset_window_position()
+            case 't':
+                self._main_window.viewport_handler.reset_window_rotation()
+            case 'z':
+                self._main_window.viewport_handler.reescale_window(Vector(1.1, 1.1, 1.0))
+            case 'c':
+                self._main_window.viewport_handler.reescale_window(Vector(0.9, 0.9, 1.0))
+            case 'y':
+                self._main_window.viewport_handler.reset_window_scale()
+            case _:
+                pass
+
     def set_mode(self, user_data, mode: ObjectType) -> None:
         '''
         Define o modo.
@@ -298,6 +329,7 @@ class EditorHandler():
 
             object_index = self._main_window.display_file_handler.objects.index(self._focus_object)
             self._main_window.display_file_handler.update_object_info(object_index)
+            self._main_window.display_file_handler.request_normalization()
 
     def rescale(self, user_data) -> None:
         '''
@@ -312,6 +344,7 @@ class EditorHandler():
 
             self._focus_object.rescale(Vector(scale_x, scale_y, scale_z))
             self.update_spin_buttons()
+            self._main_window.display_file_handler.request_normalization()
 
     def rotate(self, user_data) -> None:
         '''
@@ -324,6 +357,7 @@ class EditorHandler():
 
             self._focus_object.rotate(angle, self._rotation_anchor)
             self.update_spin_buttons()
+            self._main_window.display_file_handler.request_normalization()
 
     def change_rotation_anchor(self, user_data) -> None:
         '''
@@ -375,6 +409,7 @@ class EditorHandler():
 
             object_index = self._main_window.display_file_handler.objects.index(self._focus_object)
             self._main_window.display_file_handler.update_object_info(object_index)
+            self._main_window.display_file_handler.request_normalization()
 
     def update_scale(self, user_data) -> None:
         '''
@@ -388,6 +423,7 @@ class EditorHandler():
             diff_z = self._scale_z_button.get_value() / self._focus_object.scale.z
 
             self._focus_object.rescale(Vector(diff_x, diff_y, diff_z))
+            self._main_window.display_file_handler.request_normalization()
 
     def update_rotation(self, user_data) -> None:
         '''
@@ -401,6 +437,7 @@ class EditorHandler():
             diff_z = self._rotation_z_button.get_value() - self._focus_object.rotation.z
 
             self._focus_object.rotate(diff_z)
+            self._main_window.display_file_handler.request_normalization()
 
     def update_rotation_anchor(self, user_data) -> None:
         '''