+++ source/file_system.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8 -*-
 
 '''
-...
+Módulo para o gerenciamento de arquivos.
 '''
 
 
 class FileSystem():
     '''
+    Sistema de arquivos.
     '''
 
     def __init__(self) -> None: