+++ source/transform.py
@@ -4,7 +4,7 @@
 Módulo para as operações matemáticas.
 '''
 
-from math import cos, radians, sin
+from math import cos, radians, sin, sqrt, acos
 import numpy as np
 
 
@@ -40,9 +40,15 @@ class Vector():
 
     def __mul__(self, other):
 
+        result = None
         if isinstance(other, (float, int)):
-            return Vector(self.x * other, self.y * other, self.z * other)
-        raise NotImplementedError
+            result = Vector(self.x * other, self.y * other, self.z * other)
+        elif isinstance(other, Vector):
+            result = acos(self.dot_product(other) / (self.magnitude() * other.magnitude()))
+        else:
+            raise NotImplementedError
+
+        return result
 
     def __truediv__(self, other):
 
@@ -50,6 +56,19 @@ class Vector():
             return Vector(self.x / other, self.y / other, self.z / other)
         raise NotImplementedError
 
+    def dot_product(self, other):
+        '''
+        Produto escalar.
+        '''
+
+        return self.x * other.x + self.y * other.y + self.z * other.z
+
+    def magnitude(self) -> float:
+        '''
+        Retorna a magnitude.
+        '''
+
+        return sqrt(self.x**2 + self.y**2 + self.z**2)
 
 class Transform():
 
@@ -66,6 +85,7 @@ class Transform():
     _rotation_matrix_x: np.matrix
     _rotation_matrix_y: np.matrix
     _rotation_matrix_z: np.matrix
+    _normalization_matrix: np.matrix
 
     def __init__(self,
                  position: Vector,
@@ -75,7 +95,6 @@ class Transform():
         self._position = position
         self._rotation = rotation
         self._scale = scale
-
 
         self._translation_matrix = np.matrix([[1.0, 0.0, 0.0, self._position.x],
                                               [0.0, 1.0, 0.0, self._position.y],
@@ -111,6 +130,11 @@ class Transform():
                                              [0.0, 0.0, 1.0, 0.0],
                                              [0.0, 0.0, 0.0, 1.0]])
 
+
+        self._normalization_matrix = np.matrix([[cos_0, -sin_0, 0.0, self._position.x],
+                                                [sin_0, cos_0, 0.0, self._position.y],
+                                                [0.0, 0.0, 1.0, self._position.z],
+                                                [0.0, 0.0, 0.0, 1.0]])
 
     def __repr__(self) -> str:
         return str(f"P: {self.position}, S: {self.scale}, R: {self._rotation}")
@@ -148,9 +172,9 @@ class Transform():
         '''
 
         if anchor is None:
-            return coord - self.position
-        else:
-            return coord - anchor
+            return coord - self._position
+
+        return coord - anchor
 
     def local_to_world(self, coord: Vector, anchor: Vector = None) -> Vector:
         '''
@@ -158,17 +182,21 @@ class Transform():
         '''
 
         if anchor is None:
-            return coord + self.position
-        else:
-            return coord + anchor
+            return coord + self._position
+
+        return coord + anchor
 
     # Transformações
-    def translate(self, direction: Vector, coords: list[Vector]) -> list:
+    def translate(self,
+                  direction: Vector,
+                  coords: list[Vector],
+                  update_internal_vectors: bool = True) -> list[Vector]:
         '''
         Translação de um objeto. Retorna uma nova lista de coordenadas.
         '''
 
-        self._position += direction
+        if update_internal_vectors:
+            self._position += direction
 
         self._translation_matrix[0, 3] = direction.x
         self._translation_matrix[1, 3] = direction.y
@@ -184,14 +212,18 @@ class Transform():
 
         return new_coords
 
-    def rescale(self, scale: Vector, coords: list[Vector]) -> list:
+    def rescale(self,
+                scale: Vector,
+                coords: list[Vector],
+                update_internal_vectors: bool = True) -> list[Vector]:
         '''
         Transformação de escala.
         '''
 
-        self._scale.x *= scale.x
-        self._scale.y *= scale.y
-        self._scale.z *= scale.z
+        if update_internal_vectors:
+            self._scale.x *= scale.x
+            self._scale.y *= scale.y
+            self._scale.z *= scale.z
 
         self._scaling_matrix[0, 0] = scale.x
         self._scaling_matrix[1, 1] = scale.y
@@ -209,14 +241,19 @@ class Transform():
 
         return new_coords
 
-    def rotate(self, angle: float, coords: list[Vector], anchor: Vector = None) -> list:
+    def rotate(self,
+               angle: float,
+               coords: list[Vector],
+               anchor: Vector = None,
+               update_internal_vectors: bool = True) -> list[Vector]:
         '''
         Rotaciona o objeto em relação à um ponto.
         '''
 
         #TODO: Adicionar a rotação em múltiplos eixos
 
-        self._rotation.z = (self._rotation.z + angle) % 360
+        if update_internal_vectors:
+            self._rotation.z = (self._rotation.z + angle) % 360
 
         angle_cos = cos(radians(angle))
         angle_sin = sin(radians(angle))
@@ -226,16 +263,52 @@ class Transform():
         self._rotation_matrix_z[1, 0] = angle_sin
         self._rotation_matrix_z[1, 1] = angle_cos
 
+        corrected_coords = coords + [self._position] if update_internal_vectors else coords
         new_coords = []
 
         # Translação ponto a ponto. A posição é adicionada como um ponto
-        for coord in coords + [self._position]:
+        for coord in corrected_coords:
 
             relative_coord = self.world_to_local(coord, anchor)
             new_coord = np.matmul(self._rotation_matrix_z, [relative_coord.x, relative_coord.y, relative_coord.z, 1])
             relative_new_coord = self.local_to_world(Vector(new_coord[0, 0], new_coord[0, 1], new_coord[0, 2]), anchor)
             new_coords.append(relative_new_coord)
 
-        self._position = new_coords[-1]  # Atualiza a posição
+        if update_internal_vectors:
+            self._position = new_coords[-1]  # Atualiza a posição
+            return new_coords[:-1]  # Retorna todas as coordenadas menos a posição
 
-        return new_coords[:-1]  # Retorna todas as coordenadas menos a posição
+        return new_coords
+
+    def normalize(self,
+                  window_center: Vector,
+                  window_rotation: float,
+                  scale: Vector,
+                  coords: list[Vector]) -> list[Vector]:
+        '''
+        Normaliza as coordenadas.
+        '''
+
+        new_coords = []
+
+        self._normalization_matrix[0, 3] = -window_center.x
+        self._normalization_matrix[1, 3] = -window_center.y
+        self._normalization_matrix[2, 3] = -window_center.z
+
+        angle_cos = cos(radians(-window_rotation))
+        angle_sin = sin(radians(-window_rotation))
+
+        self._normalization_matrix[0, 0] = angle_cos
+        self._normalization_matrix[0, 1] = -angle_sin
+        self._normalization_matrix[1, 0] = angle_sin
+        self._normalization_matrix[1, 1] = angle_cos
+
+        for coord in coords:
+
+            relative_coord = self.world_to_local(coord, Vector(0.0, 0.0, 0.0))
+            new_coord = np.matmul(self._normalization_matrix, [relative_coord.x, relative_coord.y, relative_coord.z, 1])
+            relative_new_coord = self.local_to_world(Vector(new_coord[0, 0], new_coord[0, 1], new_coord[0, 2]),
+                                                     Vector(0.0, 0.0, 0.0))
+            new_coords.append(relative_new_coord)
+
+        return new_coords